Improved SparseChunkedStackArray* and adds StackArray.contentVersion
